requirements.txt
pytest>=7.0.0
boto3>=1.26.0
pyyaml>=6.0
